% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cleaned_data.R
\name{get_cleaned_data}
\alias{get_cleaned_data}
\title{Get Cleaned Data}
\usage{
get_cleaned_data()
}
\value{
A data frame with the cleaned renewable energy data.
}
\description{
This function retrieves the cleaned dataset for analysis.
}
